// hivemq-edge-frontend/src/api/__generated__/models/AdaptersList.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Adapter } from './Adapter';

export type AdaptersList = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<Adapter>;
};