// hivemq-edge-frontend/src/api/__generated__/models/DomainTagList.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { DomainTag } from './DomainTag';

export type DomainTagList = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<DomainTag>;
};