// hivemq-edge-frontend/src/api/__generated__/models/ProtocolAdaptersList.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ProtocolAdapter } from './ProtocolAdapter';

export type ProtocolAdaptersList = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<ProtocolAdapter>;
};